MOD / 3-3-1. 남은 시간 확인 / SSE 연결 완료, 메시지 테스트 중

See merge request s09-webmobile3-sub2/S09P12A203!23

+++ springboot/src/main/java/com/petmeeting/springboot/domain/Board.java
@@ -42,16 +42,15 @@ public class Board {
     private String content;
 
     @Column(name = "created_time", nullable = false)
-    private Integer createdTime;
+    private Long createdTime;
 
     @Column(name = "modified_time")
-    private Integer modifiedTime;
+    private Long modifiedTime;
 
     @Column(name = "deleted_time")
-    private Integer deletedTime;
+    private Long deletedTime;
 
     @Column(name = "view_cnt", nullable = false)
     @ColumnDefault("0")
     private Integer viewCnt;
-
 }
